+++ src/components/NationalText/index.ts
@@ -1,3 +1,5 @@
 import NationalText from './NationalText';
+import { getNationalText } from './utils';
 
+export { getNationalText };
 export default NationalText;
